+++ x8086NetEmu/Helpers/Misc/InterruptController.vb
@@ -1,4 +1,4 @@
 ﻿Public MustInherit Class InterruptController
     Implements IInterruptController
-    Public MustOverride Function GetPendingInterrupt() As Integer Implements IInterruptController.GetPendingInterrupt
+    Public MustOverride Function GetPendingInterrupt() As Byte Implements IInterruptController.GetPendingInterrupt
 End Class
